add correct sine formular

+++ lib/src/main/java/de/mario222k/wavedrawable/Wave.kt
@@ -11,7 +11,7 @@ class Wave(
     var frequency: Float,
     var x: Float,
     var y: Float,
-    var samples: Float
+    var samples: Int
 ) {
 
     fun setColor(@ColorInt color: Int) {
